statistics

+++ app/Http/Tenant/Email/Controllers/EmailController.php
@@ -92,26 +92,34 @@ class EmailController extends BaseController {
     function get()
     {
         $id = $this->request->route('id');
-        $mail = $this->email->with('attachments', 'receivers')->where('id', $id)->user()->first()->toArray();
+        $folder = $this->request->input('folder');
 
-        $to = "";
-        $cc = "";
-        foreach ($mail['receivers'] as $re) {
-            if ($re['type'] == 1)
-                $to .= $re['email'] . ';';
-            else
-                $cc .= $re['email'];
+        if($folder == 0) {
+            $mail = $this->email->with('attachments', 'receivers')->where('id', $id)->user()->first()->toArray();
+            $to = "";
+            $cc = "";
+            foreach ($mail['receivers'] as $re) {
+                if ($re['type'] == 1)
+                    $to .= $re['email'] . ';';
+                else
+                    $cc .= $re['email'];
+            }
+
+            foreach ($mail['attachments'] as &$at) {
+                unset($at['id']);
+                unset($at['email_id']);
+            }
+
+            unset($mail['receivers']);
+            $mail['to'] = $to;
+            $mail['cc'] = $cc;
         }
-
-        foreach ($mail['attachments'] as &$at) {
-            unset($at['id']);
-            unset($at['email_id']);
+        else {
+            $mail = $this->incomingEmail->where('id', $id)->user()->first()->toArray();
+            $mail['message'] = $mail['body_html'];
+            $mail['to'] = $mail['from_email'];
+            $mail['cc'] = '';
         }
-
-
-        unset($mail['receivers']);
-        $mail['to'] = $to;
-        $mail['cc'] = $cc;
 
         $display_name = current_user()->display_name;
         $email = current_user()->smtp->email;
@@ -188,20 +196,31 @@ class EmailController extends BaseController {
     {
         if ($this->request->ajax()) {
             $id = $this->request->route('id');
-            $email = $this->email->where('id', $id)->user()->first();
-            if (!empty($email)) {
-                if ($email->delete()) {
+            $folder = $this->request->input('folder');
 
-                    $this->attachment->where('email_id', $id)->delete();
-                    $this->receiver->where('email_id', $id)->delete();
+            if($folder == 0) {
+                $email = $this->email->where('id', $id)->user()->first();
+                if (!empty($email)) {
+                    if ($email->delete()) {
 
-                    return $this->success(['message' => 'Email deleted Successfully']);
+                        $this->attachment->where('email_id', $id)->delete();
+                        $this->receiver->where('email_id', $id)->delete();
+
+                        return $this->success(['message' => 'Email deleted Successfully!']);
+                    }
+                }
+            }
+            else{
+                $email = $this->incomingEmail->where('id', $id)->user()->first();
+                if (!empty($email)) {
+                    if ($email->delete()) {
+                        return $this->success(['message' => 'Email deleted Successfully!']);
+                    }
                 }
             }
 
             return $this->fail(['error' => 'Something went wrong. Please try again later']);
         }
-
     }
